Add files via upload
Was my final project for CPTS_484 where we had to develop a front and backend to a website and I scored an 84% on it (42/50) we had around an hour or two to develop and set up a front and backend and were able to use code snippets from previous assignments and also grab anything needed from sources for overall appearence of the project. i.e. button interactions, drop down menu's, ect. We were building upon and improving our midterm assignments to meet the finals. The screenshot is the final image from the project.