Add files via upload
This is my midterm where we were mimicing a page that the professor provided an image of. In the hour provided I spent time working on creating a copy that was around a 77% for the final grade(23.25/30) as I was missing or lacking on certain parts of the assignment. We were allowed to look around online for syntax and look through our notes and books.